AI_HR/backend/requirements.txt
fastapi
uvicorn[standard]
websockets
vosk
numpy
openai
python-dotenv
python-multipart
pymupdf4llm
python-docx
